requirements.txt
pyzmq
geventhttpclient
fabric
geventhttpclient
sphinx
